feat: add primer
