+++ src/style.scss
@@ -1,3 +1,7 @@
+@import "@primer/css/core/index.scss";
+@import "@primer/css/product/index.scss";
+@import "@primer/css/marketing/index.scss";
+
 * {
   padding: 0;
   margin: 0;
