+++ src/view-component/Counter/counter.scss
@@ -10,11 +10,5 @@
 
 .button {
   font-family: serif;
-  padding: 2em;
-  font-size: 20px;
-  border-radius: 3px;
-  &:hover {
-    cursor: pointer;
-  }
 }
