+++ src/view-component/index.ts
@@ -1,5 +1,5 @@
 import * as Counter from "./Counter/Counter";
 export { Counter };
 
-import * as CounterPage from "./CounterPage/CounterPage";
-export { CounterPage };
+import * as Page from "./Page/Page";
+export { Page };
